% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yogiseq.R
\name{char.at}
\alias{char.at}
\title{return a character froma string}
\usage{
char.at(str, i)
}
\arguments{
\item{str}{input string}

\item{i}{the character index}
}
\value{
the character
}
\description{
return's the i'th character in a string
}
