% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suffixTrees.R
\name{commonPrefix}
\alias{commonPrefix}
\title{Determine the length of the a common prefix between
two strings. For example: Banana and Band have a common
prefix of length 3.}
\usage{
commonPrefix(x, y, wildcard = NULL)
}
\arguments{
\item{x}{first string}

\item{y}{second string}

\item{wildcard}{a wildcard character that always counts as a match
or NULL to disable. Default: NULL}
}
\value{
the length of the common prefix
}
\description{
Determine the length of the a common prefix between
two strings. For example: Banana and Band have a common
prefix of length 3.
}
