% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yogiseq.R
\name{new.bc.matcher}
\alias{new.bc.matcher}
\title{Create new barcode matcher}
\usage{
new.bc.matcher(lib, errCutoff = 2, strictMode = FALSE)
}
\arguments{
\item{lib}{library of barcode sequences to match against}
}
\value{
a new barcode matcher objects with the method:
   findMatches(queries)
}
\description{
Create new barcode matcher
}
