% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yogiseq.R
\name{new.fastq.parser}
\alias{new.fastq.parser}
\title{creates a new fastq parser object}
\usage{
new.fastq.parser(con)
}
\arguments{
\item{file}{or connection}
}
\value{
new fastq parser with method 
    <code>parseNext(n=10,ignore.quality=FALSE)</code>
}
\description{
creates a new fastq parser object
}
