% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yogiseq.R
\name{new.kmer.search}
\alias{new.kmer.search}
\title{K-mer search}
\usage{
new.kmer.search(k = 5)
}
\arguments{
\item{k}{length of k-mers}
}
\value{
the k-mer search object with methods:
   <code>build.index(fasta.file)</code>
   <code>load.index(index.file)</code>
   <code>search(queries,min.hits=3,max.d=Inf)</code>
}
\description{
K-mer search
}
