% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samStreamer.R
\name{new.sam.streamer}
\alias{new.sam.streamer}
\title{construct new SAM streamer object}
\usage{
new.sam.streamer(sam.file, chunkSize = 100)
}
\arguments{
\item{sam.file}{the input sam file or connection object}

\item{chunkSize}{how many lines to process at once}
}
\value{
a sam.streamer object with methods to access the content
}
\description{
can be used to stream
chunks of SAM files to be processed iteratively. For each chunk,
the individual columns as well as parsed flags, MDZ and CIGAR strings
are made available
}
