% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yogiseq.R
\name{new.sequence}
\alias{new.sequence}
\title{constructor to create a new sequence object}
\usage{
new.sequence(sequence, qual = NULL, id = NULL)
}
\arguments{
\item{sequence}{the sequence}

\item{qual}{the quality string}

\item{id}{an identifier for the sequence}
}
\value{
the sequence object
}
\description{
creates new yogiseq object
}
