% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samStreamer.R
\name{parseMDZ}
\alias{parseMDZ}
\title{Tokenize MDZ string, breaking it down into a list of number and string elemetns
E.g. "MD:Z:3A5^AC8" would be broken down into list(3,"A",5,"^AC",8)}
\usage{
parseMDZ(mdz)
}
\description{
Tokenize MDZ string, breaking it down into a list of number and string elemetns
E.g. "MD:Z:3A5^AC8" would be broken down into list(3,"A",5,"^AC",8)
}
