% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yogiseq.R
\name{readFASTA}
\alias{readFASTA}
\title{read sequences from fasta file}
\usage{
readFASTA(con)
}
\arguments{
\item{con}{input file or connection}
}
\value{
the sequences
}
\description{
read sequences from fasta file
}
