% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yogiseq.R
\name{reverseComplement}
\alias{reverseComplement}
\title{reverse complement of sequence}
\usage{
reverseComplement(seq)
}
\arguments{
\item{seq}{the yogiseq object}
}
\value{
the reverse complement of the sequence
}
\description{
reverse complement of sequence
}
