% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suffixTrees.R
\name{searchSuffixTree}
\alias{searchSuffixTree}
\title{Search for a query string within a suffix tree
with or without wildcard characters.}
\usage{
searchSuffixTree(tree, query, wildcard = NULL)
}
\arguments{
\item{tree}{the suffix tree}

\item{query}{the query string}

\item{wildcard}{A wildcard character that will match anything
or NULL to disable. Default: NULL}
}
\value{
An integer vector indicating the start positions of
        any matches
}
\description{
Search for a query string within a suffix tree
with or without wildcard characters.
}
