% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yogiseq.R
\name{subseq}
\alias{subseq}
\title{subsequence}
\usage{
subseq(s, from, to)
}
\arguments{
\item{from}{from index}

\item{to}{end index}

\item{seq}{the yogiseq object}
}
\value{
given subsequece
}
\description{
subsequence
}
