% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suffixTrees.R
\name{suffixTree}
\alias{suffixTree}
\title{Construct a suffix tree for a given string.}
\usage{
suffixTree(str)
}
\arguments{
\item{str}{the input string}
}
\value{
the suffix tree (implemented as list of lists)
}
\description{
Construct a suffix tree for a given string.
}
