% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yogiseq.R
\name{to.char.array}
\alias{to.char.array}
\title{string to character vector}
\usage{
to.char.array(str)
}
\arguments{
\item{str}{input string}
}
\value{
a vector of single characters
}
\description{
Turns a string into a vector of single characters
}
