% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samStreamer.R
\name{varcallCandidates}
\alias{varcallCandidates}
\title{Find variant call candidates from a sam file line}
\usage{
varcallCandidates(flagV, mdzStr, cigarStr, startPos, readSeq, readQual)
}
\arguments{
\item{flagV}{vector of flags (from samStreamer)}

\item{mdzStr}{MD:Z string from sam}

\item{cigarStr}{CIGAR string from sam}

\item{startPos}{alignment start position from sam}

\item{readSeq}{read sequence from sam}

\item{readQual}{quality string from sam}
}
\value{
a table of all variant base calls
}
\description{
Interpret the CIGAR and MDZ strings to construct a 
table of variant base calls
}
