% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yogiseq.R
\name{writeFASTA}
\alias{writeFASTA}
\title{write sequences to fasta file}
\usage{
writeFASTA(con, seqs)
}
\arguments{
\item{con}{output connection}

\item{seqs}{list of yogiseq objects}
}
\value{
given subsequece
}
\description{
write sequences to fasta file
}
