% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yogiseq.R
\name{writeFASTQ}
\alias{writeFASTQ}
\title{write sequences to fastq file}
\usage{
writeFASTQ(con, seqs)
}
\arguments{
\item{con}{file or connection}

\item{seqs}{list of sequences}
}
\description{
write sequences to fastq file
}
